refactor(language-server): move class code completion tests to file
The tests did not change except for one test that uses a new utility
function to get the range for the additionalTextInserts:
"will insert the namespace when selecting completion for class in inner tag and namespace is not defined"

Utility functions moved to their own file.
Changes to utility functions:
* Added assertion that text edits don't overlap (assertRangesDoNotOverlap)
* Added model parameter to getSuggestions (since it's no longer in a
  closure)
* Added handling of additionalTextInserts ranges in getSuggestions
* Added function getRanges
* Extracted function getXmlSnippetDocument from getSuggestions
